Update user profile page
- Separate password reset and authentication tokens into tabs
- Adds modal for token creation
- Adds the ability to view and copy token value upon creation
