+++ package.json
@@ -18,6 +18,7 @@
     "bulma": "^0.7.1",
     "bulma-tooltip": "^2.0.2",
     "bulmaswatch": "^0.7.2",
+    "clipboard": "^2.0.4",
     "core-js": "^2.6.5",
     "d3": "^4.10.2",
     "d3-relationshipgraph": "^3.0.2",
