+++ src/styles/components/_user-authentication-tokens.scss
@@ -1,98 +1,102 @@
 .user-authentication-tokens {
 	.header {
 		display: flex;
-		justify-content: space-between;
 		align-items: center;
 		margin-bottom: 20px;
 
-		.subtitle {
-			margin-bottom: 0;
+		.button {
+			border-radius: 4px;
 		}
 
-		button {
+		.control {
+			margin-right:20px;
+		}
+
+		input {
+			background-color: $dark;
+			border-color: $grey;
 			border-radius: 4px;
+			color: #bdcbdb;
+			margin-right: 15px;
+			min-width: 350px;
+
+			&::placeholder {
+				color: #bdcbdb;
+			}
 		}
 	}
 
-	.card-content {
-		padding: 0;
+	.table thead {
+		th {
+			border-width: 0 0 1px 0;
 
-		thead {
-			th {
-				background-color: $grey-darker;
-				border-width: 0 0 1px 0;
+			&:first-of-type {
+				border-top-left-radius: 5px;
+			}
 
-				&:first-of-type {
-					border-top-left-radius: 5px;
-				}
-
-				&:last-of-type {
-					border-top-right-radius: 5px;
-				}
+			&:last-of-type {
+				border-top-right-radius: 5px;
 			}
 		}
+	}
 
-		tfoot {
-			th {
-				background-color: $grey-darker;
-				border-top-width: 1px;
+	tfoot {
+		th {
+			border-top-width: 1px;
 
-				&:first-of-type {
-					border-bottom-left-radius: 5px;
-				}
+			&:first-of-type {
+				border-bottom-left-radius: 5px;
+			}
 
-				&:last-of-type {
-					border-bottom-right-radius: 5px;
-				}
+			&:last-of-type {
+				border-bottom-right-radius: 5px;
 			}
 		}
+	}
 
-		.token {
-			background-color: darken($grey-darker, 4%);
-			transition: background-color 0.3s ease-in-out;
-
-			&.create {
-				input {
-					background-color: $grey-dark;
-					color: $grey-lighter;
-
-					&::placeholder {
-						color: $grey-lighter;
-					}
-				}
-
-				td {
-					.button {
-						border-top-right-radius: 4px;
-						border-bottom-right-radius: 4px;
-					}
-				}
-			}
-
-			.has-success-background {
-				background-color: $green;
-				transition: background-color 0.3s ease-in-out;
-			}
-
-			.delete-token {
-				&:hover {
-					color: $red;
-					cursor: pointer;
-				}
-			}
-
-			td {
-				padding: 14px;
-				white-space: nowrap;
-
-				&:last-of-type {
-					text-align: right;
-				}
-
-				&:first-of-type {
-					width: 22px;
-				}
-			}
+	.delete-token {
+		&:hover {
+			color: $red;
+			cursor: pointer;
 		}
+	}
+
+	td {
+		padding: 14px;
+		white-space: nowrap;
+
+		&:last-of-type {
+			text-align: right;
+		}
+
+		&:first-of-type {
+			width: 22px;
+		}
+	}
+
+	.empty-state {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		justify-content: center;
+		margin-top: 80px;
+
+		.empty-state-heading {
+			color: $white;
+			font-size: 24px;
+			font-weight: bold;
+		}
+
+		.button {
+			font-weight: bold;
+			font-size: 17px;
+			margin: 20px 0 40px;
+			height: 44px;
+			min-width: 180px;
+		}
+	}
+
+	.no-search-results {
+		margin-top: 100px;
 	}
 }
