+++ src/styles/components/_user-profile.scss
@@ -1,19 +1,22 @@
 .user-profile {
-	section.content {
-		margin-top: 0.75rem;
+	.tabs {
+		margin-bottom: 20px;
 
-		.message-header {
-			i {
-				margin-right: 10px;
+		li {
+			&:first-of-type {
+				a {
+					border-top-left-radius: 4px;
+					border-bottom-left-radius: 4px;
+				}
+
 			}
 
-			p {
-				margin-bottom: 0;
+			&:last-of-type {
+				a {
+					border-top-right-radius: 4px;
+					border-bottom-right-radius: 4px;
+				}
 			}
-		}
-
-		.button.save {
-			margin-top: 0.75rem;
 		}
 	}
