+++ src/styles/main.scss
@@ -26,6 +26,7 @@
 @import './components/app';
 @import './components/authentication-tokens';
 @import './components/base-modal';
+@import './components/create-token-modal';
 @import './components/data-center-browser';
 @import './components/device-modal';
 @import './components/edit-layout-modal';
